Ensure git is installed for all reactive openstack charms
This is due to bug 1767328, which is due to setuptools_scm using
the 'git' command to find the version number when charms.openstack
is installed (via the wheelhouse.txt) from git+https://...

This only happens on trusty as xenial+ already have git installed.
The script only tries to install git once and then uses a local
flag file to stop it trying thereafter on every hook.

Change-Id: I8a7daa0fff3d5bc7b11f70cde246975987c7cb66
Closes-Bug: #1767328