Backout actions.yaml as a tactical fix.
The current implementation forces a charm to define its own
pause/resume actions and charm lint will fail if they do not. This
breaks charms in an unpleasant way so instead work is ongoing
to make pause/resume code standalone in this layer. For the moment
backout actions to unbreak charms.

Change-Id: I04f313097f632d7fa89e1734b966d12bda21656c