Update layer wheelhouse with common dependencies
Add empty series list to metadata.yaml to get around charm-tools
now mandating a series list to be present.

This list will always be overridden by charms but building the
layer is part of our gate and we need to unblock it.

Update tox basepython, the build will no longer succeed in the gate
with py27.

Change-Id: If81441d5fb0ed3b8819c4a2814dc20c5e47a8eff