Add a 'is-update-status-hook' flag to the layer
This flag is to enable charms.openstack / layer-openstack charms to have
a way to determine if the current hook is update-status from a flag.  By
being a flag, it means that charm reactive handler code doesn't have to
check for the the hook in every handlers, but can declare "don't run
this if it's an update-status hook".  The flag is guaranteed to not be
set during a not update-status hook (in that, if it is set, then it gets
cleared and the handlers can therefore run), or to be set during an
update-status hook.  To speed things up, the code also removes the flag
at the end of the update-status handler.

The intended use of this flag is to guard 'expensive' handlers from
being run during update-status.

Change-Id: I7ca97db646ada95d7f5541ca4e2cf14013c0d7a6