Do not execute certificate handlers in update-status hook
The certificate handler code does a bit of work and should not
run during the update-status hook.

As documented in the charms.reactive dispatch [0] function
handlers registered with the @hook decorator will execute before
other handlers.  So we can rely on the 'is-update-status-hook'
flag being consistently set when in update-status hook before
other handlers are evaluated for dispatch.

0: https://github.com/juju-solutions/charms.reactive/blob/3d27f3d24a2dcdc81f2a2299355b110c5c68d48c/charms/reactive/bus.py#L325

Closes-Bug: #1954748
Change-Id: I4a3aa544f98049c83db576f95de826038e8e1afc