fix: actions need hooks/
Some charm builds will put additional functionality into a subdirectory
of hooks/ (eg designate-bind creates a hooks/relations directory.)

However, when running an action, those are run from "actions/pause" not
as "hooks/config-changed". Which means that they cannot find the
"relations" directory to import it.

We could:
a) symlink all the extra things from actions/* to point at hooks/*
b) include hooks as part of the import path
c) change so that all of those extra libraries are put into somewhere in
'lib' instead of somewhere in hooks/*.

This patch chooses (b).

This should fix:
https://bugs.launchpad.net/charm-designate-bind/+bug/2107591

once it is rebuilt with this version of layer openstack.